Enable FIPS mode for IPA if system is in FIPS mode
If FIPS is enabled in the hosts we should also run IPA in FIPS mode.
It is possible to enable FIPS directly at kernel level using the
fips option, determining the FIPS status for example from
the cryptographic module and specifically the
/proc/sys/crypto/fips_enabled file; if the file contains 1 then
the system is in FIPS mode, if it contains 0 the FIPS algorithms
are disabled.
Therefore the value of the fips kernel option is 0 (default)
if FIPS is disabled, or 1 if enabled.

Signed-off-by: Riccardo Pittau <elfosardo@gmail.com>