Do not use /tmp as a source directory for the image
Unfortunately, a lot of code in Ironic ignores the tempdir
configuration. Even worse, the default value for this option relies on
having a writable /tmp, which makes the current approach not functional
without mounting an emptyDir to /tmp. But if an emptyDir is mounted to
/tmp, we lose access to all the template files stored there.

This change uses a new directory /templates as a source for files that
needs copying/templating in runtime.

Also fix usage of /var/tmp in httpd.conf.

Document the requirements.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>
(cherry picked from commit 8f8b12b141cfec87ee73f80fd15e091d00f6e1c1)