METAL-1695: OKD: install Ironic from the OpenShift fork
Using PyPI releases is risky since ironic-image is sometimes changed
faster and relies on recent commits. Use the OpenShift fork of Ironic
instead, which is guaranteed to work with ironic-image.

As an aside, this still assumes using the main branch for all OKD
branches. We need to discuss this part separately.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>